replaced theme switcher with +button
