+++ src/components/HeaderDashboard/HeaderDashboard.scss
@@ -48,9 +48,8 @@
   background: transparent;
 }
 .dark-mode-toggle {
-  position: absolute;
-  top: 25px;
-  right: 18px;
+  margin-left: 138px;
+  margin-top: 90px;
 }
 #goal-suggestion-btn, #goal-suggestion-btn-dark {
     position: absolute;
