+++ src/components/HeaderDashboard/MainHeaderDashboard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useRecoilState, useSetRecoilState } from "recoil";
+import { useRecoilValue, useSetRecoilState } from "recoil";
 import { useNavigate } from "react-router-dom";
 
 import { darkModeState } from "@store";
@@ -9,24 +9,19 @@ import mainAvatarDark from "@assets/images/mainAvatarDark.svg";
 import ZinZenTextLight from "@assets/images/LogoTextLight.svg";
 import ZinZenTextDark from "@assets/images/LogoTextDark.svg";
 import { displaySidebar } from "@src/store/SidebarState";
-import Sidebar from "@components/Sidebar";
-import DarkModeToggle from "react-dark-mode-toggle";
+import { displayAddGoalOptions } from "@src/store/GoalsState";
+import plus from "@assets/images/plus.svg";
 
+import Sidebar from "@components/Sidebar";
 import "@translations/i18n";
 import "./HeaderDashboard.scss";
 
 export const MainHeaderDashboard = () => {
   const navigate = useNavigate();
   const setShowSidebar = useSetRecoilState(displaySidebar);
-  const [darkModeStatus, setDarkModeStatus] = useRecoilState(darkModeState);
-  const toggleTheme = () => {
-    setDarkModeStatus(!darkModeStatus);
-    if (darkModeStatus) {
-      localStorage.setItem("theme", "light");
-    } else {
-      localStorage.setItem("theme", "dark");
-    }
-  };
+  const darkModeStatus = useRecoilValue(darkModeState);
+  const setShowAddGoalOptions = useSetRecoilState(displayAddGoalOptions);
+
   return (
     <>
       <img
@@ -46,12 +41,16 @@ export const MainHeaderDashboard = () => {
           navigate("/MyGoals");
         }}
       />
-      <DarkModeToggle
-        onChange={toggleTheme}
-        checked={darkModeStatus}
-        size={60}
-        className="dark-mode-toggle"
-      />
+      <button
+        type="button"
+        id={`goal-suggestion-btn${darkModeStatus ? "-dark" : ""}`}
+        onClick={async () => {
+          navigate("/MyGoals");
+          setShowAddGoalOptions(true);
+        }}
+      >
+        <img alt="save changes" src={plus} />
+      </button>
       <Sidebar />
     </>
